Update default featured image to use dynamic values.
Chang how default featured image was being set. Update the function so it dynamically searches for "Aloha Friday" category and gets the ID. It pulls the attachment ID for the featured image from the default featured image selected on the settings page. It had been using hard-coded values to check for the correct category and assign the featured image.